modify test

+++ ivy_tests/test_ivy/test_frontends/test_torch/test_reduction_ops.py
@@ -610,7 +610,7 @@ def test_torch_std_mean(
 @handle_frontend_test(
     fn_tree="torch.count_nonzero",
     dtype_input_axis=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("numeric"),
+        available_dtypes=helpers.get_dtypes("float"),
         force_int_axis=True,
         min_num_dims=1,
         min_axis=-1,
